rename module for crate private extensions
